Use websocket instead of inindexer in gpt4o-indexer
