+++ crates/gpt4o-indexer/Cargo.toml
@@ -1,15 +1,12 @@
 [package]
-name = "indexer"
+name = "gpt4o-indexer"
 version = "0.1.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
 
 [dependencies]
 dotenvy = "0.15.7"
-inindexer = { version = "0.1.0-alpha.3", features = [ "fastnear-data-server" ] }
 tokio = { version = "1.38.0", features = [ "rt-multi-thread", "macros" ] }
-async-trait = "0.1.80"
-simple_logger = "5.0.0"
 log = "0.4.21"
 near-api = "0.3.0"
 serde = { version = "1.0.203", features = [ "derive" ] }
@@ -19,3 +16,8 @@ reqwest = { version = "0.12.4", features = [ "json" ] }
 near-gas = "0.3.0"
 near-token = "0.3.0"
 near-primitives = "0.26.0"
+env_logger = "0.11.1"
+inevents-websocket-client = "1.0.0"
+intear-events = { git = "https://github.com/INTEARnear/inevents" }
+json-filter = "0.1.0"
+futures = "0.3.30"
