+++ crates/gpt4o-indexer/README.md
